Enhance interactive mode: add command-line argument support for interactive mode in test files

+++ tests/test_coco.cpp
@@ -24,8 +24,12 @@
 #include <thread>
 #endif
 
-int main()
+int main([[maybe_unused]] int argc, [[maybe_unused]] char *argv[])
 {
+    bool interactive = false;
+    if (argc > 1 && std::string(argv[1]) == "--interactive")
+        interactive = true;
+
 #ifdef BUILD_MONGODB
     mongocxx::instance inst{}; // This should be done only once.
     coco::mongo_db db;
@@ -86,18 +90,21 @@ int main()
     auto &s_itm = cc.create_item(s_tp);
     cc.set_value(s_itm, json::json{{"parent", ch_itm.get_id().c_str()}});
 
-    std::string user_input;
-    do
+    if (interactive)
     {
-        std::cout << "Enter a command (d to drop the database, q to quit): ";
-        std::getline(std::cin, user_input);
-        if (user_input == "d")
-            db.drop();
+        std::string user_input;
+        do
+        {
+            std::cout << "Enter a command (d to drop the database, q to quit): ";
+            std::getline(std::cin, user_input);
+            if (user_input == "d")
+                db.drop();
 #ifdef BUILD_LLM
-        else
-            llm.understand(itm_0, user_input);
+            else
+                llm.understand(itm_0, user_input);
 #endif
-    } while (user_input != "d" && user_input != "q");
+        } while (user_input != "d" && user_input != "q");
+    }
 
 #ifdef BUILD_SERVER
     srv.stop();
